[FIX] 중복채점 방지 및 소수점 출력 제거

+++ src/main/java/swm_nm/morandi/test/dto/TestTypeDto.java
@@ -14,6 +14,6 @@ public class TestTypeDto {
     private Integer problemCount;
     private DifficultyLevel startDifficulty;
     private DifficultyLevel endDifficulty;
-    private Double averageCorrectAnswerRate;
+    private Long averageCorrectAnswerRate;
     private Integer numberOfTestTrial;
 }
